Refactor: Separate formatting logic in QueryParameterStackView
