+++ iOS/Targets/Cherrybnb/Sources/Search_Date/View/QueryParameterStackView.swift
@@ -43,6 +43,7 @@ class QueryParameterStackView: UIStackView {
     // MARK: - Properties
 
     var queryParameter: QueryParameter?
+    var formatter = QueryParameterFormatter()
 
     // MARK: - Initializers
 
@@ -81,52 +82,11 @@ class QueryParameterStackView: UIStackView {
     }
 
     private func setSubviews() {
-        for (keyText, valueText) in toString(queryParameter) {
+        for (keyText, valueText) in formatter.toString(queryParameter) {
             let parameterRow = parameterRow(keyText, valueText)
             addArrangedSubview(parameterRow)
         }
     }
 
-    private func toString(_ queryParameter: QueryParameter?) -> [(String, String)] {
-        let locationString = toString(queryParameter?.place)
-        let dateString = toString(queryParameter?.dateRange)
-
-        // TODO: Formatting Logic for price & headCount
-        let priceString = toString(queryParameter?.priceRange)
-        let headCountString = ""
-
-        return [
-            ("위치", locationString),
-            ("체크인/체크아웃", dateString),
-            ("요금", priceString),
-            ("인원", headCountString)
-        ]
-    }
-
-    private func toString(_ place: Place?) -> String {
-        return place?.name ?? ""
-    }
-
-    private func toString(_ range: (Date?, Date?)?) -> String {
-        guard let range = range else { return "" }
-
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "MMM d일"
-        dateFormatter.locale = Locale(identifier: "ko")
-
-        switch range {
-        case (.some(let checkIn), .some(let checkOut)):
-            return dateFormatter.string(from: checkIn) + " - " + dateFormatter.string(from: checkOut)
-        case (.some(let checkIn), _):
-            return dateFormatter.string(from: checkIn)
-        default: return ""
-        }
-    }
-    
-    private func toString(_ priceRange: (Price?, Price?)?) -> String {
-        guard let priceRange = priceRange else { return "" }
-        guard let min = priceRange.0, let max = priceRange.1 else { return ""}
-        return "\(min)₩-\(max)₩"
-    }
     
 }
